huge refactoring 2.0
