+++ pwt-tests/package.json
@@ -8,7 +8,7 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@playwright/test": "^1.44.1",
+    "@playwright/test": "^1.44.0",
     "@types/node": "^20.12.12"
   }
 }
