+++ pwt-tests/playwright.config.ts
@@ -1,4 +1,3 @@
-
 import { defineConfig, devices } from '@playwright/test';
 
 /**
