+++ pwt-tests/pwt-tests/login.spec.ts
@@ -1,4 +1,3 @@
-
 import { test, expect } from '@playwright/test';
 
 test.describe('Login Form', () => {
