+++ pwt-tests/pwt-tests/prices-sorting.spec.ts
@@ -1,4 +1,3 @@
-
 import { test, expect } from '@playwright/test';
 
 test('Проверка сортировки карточек по стоимости', async ({ page }) => {
