+++ src/components/offer/nearby-offers/nearby-offers.tsx
@@ -1,5 +1,5 @@
-import OffersList from '../offer/offers-list/offers-list';
-import { Offer } from '../../types/offer';
+import OffersList from '../offers-list/offers-list';
+import { Offer } from '../../../types/offer';
 
 type NearbyOffersProps = {
   nearPlaces: Offer[];
