+++ src/components/offer/offers-container/offers-container.tsx
@@ -1,6 +1,6 @@
 import { Offer } from '../../../types/offer.ts';
-import EmptyOffers from '../../empty-offers/empty-offers.tsx';
-import AvailableOffers from '../../available-offers/available-offers.tsx';
+import EmptyOffers from '../empty-offers/empty-offers.tsx';
+import AvailableOffers from '../available-offers/available-offers.tsx';
 
 interface OffersContainerProps {
   offers: Offer[];
